Remove loading spinners and implement memory as beta feature
- Remove all loading spinner components from ContentManagementTab and GroupsManagementTab
- Replace spinner loading states with static text or simple loading messages
- Maintain disabled button states for UX while removing visual spinners

- Implement memory (mem0ry) as beta feature across full stack:
  - Add memory to frontend beta features configuration and Labor tab
  - Update IntelligenceTab to use beta feature system instead of dev environment checks
  - Add memory support to backend beta features API and database schema
  - Create Supabase migration to add memory column to profiles table

- Fix memories.map error in IntelligenceTab with comprehensive array safety checks
- Add new template and content management utilities for enhanced functionality
