+++ gruenerator_frontend/src/features/templates/canva/components/CanvaAssetsPanel.jsx
@@ -1,0 +1,132 @@
+import React, { useState, useCallback } from 'react';
+import { HiPhotograph, HiExternalLink } from 'react-icons/hi';
+import AssetPackageCard from './AssetPackageCard';
+import * as canvaUtils from '../../../../components/utils/canvaUtils';
+
+/**
+ * Canva Assets Panel Component
+ * Displays available asset packages and handles import functionality
+ */
+const CanvaAssetsPanel = ({ 
+    canvaConnected, 
+    canvaLoading, 
+    onCanvaLogin, 
+    onSuccessMessage, 
+    onErrorMessage
+}) => {
+    const [importingPackage, setImportingPackage] = useState(null);
+    const [importProgress, setImportProgress] = useState('');
+    const [importedPackages, setImportedPackages] = useState(canvaUtils.getImportedPackages());
+
+    // Handle package import
+    const handleImportPackage = useCallback(async (packageId) => {
+        if (importingPackage) return; // Prevent multiple imports
+
+        setImportingPackage(packageId);
+        setImportProgress('Starte Import...');
+
+        try {
+            const result = await canvaUtils.importAssetPackage(
+                packageId,
+                (progressMsg) => setImportProgress(progressMsg),
+                (result) => {
+                    canvaUtils.markPackageAsImported(packageId);
+                    setImportedPackages(canvaUtils.getImportedPackages());
+                    onSuccessMessage(`Asset Package "${result.packageName}" wurde erfolgreich importiert! ${result.importedAssets} Assets hinzugefügt.`);
+                },
+                onErrorMessage
+            );
+        } catch (error) {
+            console.error('[CanvaAssetsPanel] Import failed:', error);
+        } finally {
+            setImportingPackage(null);
+            setImportProgress('');
+        }
+    }, [importingPackage, onSuccessMessage, onErrorMessage]);
+
+    if (!canvaConnected) {
+        return (
+            <div className="profile-card">
+                <div className="profile-card-header">
+                    <h3>Asset-Pakete</h3>
+                </div>
+                <div className="profile-card-content">
+                    <div className="empty-state">
+                        <HiPhotograph size={48} style={{ color: 'var(--font-color-muted)' }} />
+                        <p>Verbinde dein Canva-Konto, um Asset-Pakete zu verwenden.</p>
+                        <button
+                            type="button"
+                            className="btn-primary"
+                            onClick={onCanvaLogin}
+                            disabled={canvaLoading}
+                        >
+                            <HiExternalLink className="icon" />
+                            {canvaLoading ? 'Verbinde...' : 'Mit Canva verbinden'}
+                        </button>
+                    </div>
+                </div>
+            </div>
+        );
+    }
+
+    return (
+        <div className="profile-card">
+            <div className="profile-card-header">
+                <h3>Asset-Pakete</h3>
+                <p style={{ 
+                    fontSize: '0.9rem', 
+                    color: 'var(--font-color-muted)', 
+                    marginTop: 'var(--spacing-small)' 
+                }}>
+                    Wähle vorgefertigte Asset-Sammlungen aus, um sie zu deinem Canva-Konto hinzuzufügen.
+                </p>
+            </div>
+            <div className="profile-card-content">
+                {importingPackage && (
+                    <div className="import-progress" style={{ 
+                        marginBottom: 'var(--spacing-large)', 
+                        padding: 'var(--spacing-medium)', 
+                        backgroundColor: 'var(--background-color-secondary)', 
+                        borderRadius: 'var(--border-radius)' 
+                    }}>
+                        <div style={{ 
+                            display: 'flex', 
+                            alignItems: 'center', 
+                            gap: 'var(--spacing-small)' 
+                        }}>
+                            <div className="spinner" style={{ width: '20px', height: '20px' }}></div>
+                            <span>Importiere Asset-Paket...</span>
+                        </div>
+                        {importProgress && (
+                            <div style={{ 
+                                marginTop: 'var(--spacing-small)', 
+                                fontSize: '0.9rem', 
+                                color: 'var(--font-color-muted)' 
+                            }}>
+                                {importProgress}
+                            </div>
+                        )}
+                    </div>
+                )}
+
+                <div className="asset-packages-grid" style={{
+                    display: 'grid',
+                    gridTemplateColumns: 'repeat(auto-fill, minmax(300px, 1fr))',
+                    gap: 'var(--spacing-large)'
+                }}>
+                    {canvaUtils.ASSET_PACKAGES.map(pkg => (
+                        <AssetPackageCard
+                            key={pkg.id}
+                            package={pkg}
+                            isImported={importedPackages.includes(pkg.id)}
+                            isImporting={importingPackage === pkg.id}
+                            onImport={() => handleImportPackage(pkg.id)}
+                        />
+                    ))}
+                </div>
+            </div>
+        </div>
+    );
+};
+
+export default CanvaAssetsPanel;
